* master: (215 commits)
  Marking testShouldLogPutTags test incomplete.
  Updating end-to-end test to match change in title.
  Updating guest article test to include newly added $state.
  Updated title tag to be dynamic depending on $state data.
  Implementing public facing meta tags in <head> for articles.
  Applying style patch.
  Various updates to search. Now base64 encoding all query params when searching.
  Set docker-compose version to 1.3.1
  Fixing mistakes made in previous commit.
  Applying style patch.
  Various optimizations.
  Updating test to use new getFactory pattern.
  Adding test for pagination service's complex query function.
  Removing article listing's admin test (skipping for now).
  Added tests for subsection mocks
  Apply patch.
  Added tests for section content types
  Removing model/collection check from getWithNested in ApiController and writing tests for paginated simple/complex search.
  Applied fixes from StyleCI
  Fixed test and misnamed attribute for sections_display
  ...

# Conflicts:
#	api/app/Http/Controllers/EntityController.php
#	api/database/factories/ModelFactory.php

+++ api/app/Http/Controllers/EntityController.php
@@ -10,7 +10,9 @@
 
 namespace App\Http\Controllers;
 
+use App\Extensions\Controller\AuthorizesRequestsTrait;
 use Illuminate\Http\Request;
+use Illuminate\Support\Str;
 use Spira\Model\Model\BaseModel;
 use Elasticquent\ElasticquentTrait;
 use Spira\Model\Collection\Collection;
@@ -23,7 +25,7 @@ use Symfony\Component\HttpKernel\Exception\NotFoundHttpException;
 
 abstract class EntityController extends ApiController
 {
-    use RequestValidationTrait;
+    use RequestValidationTrait, AuthorizesRequestsTrait;
 
     /**
      * @var BaseModel
@@ -46,6 +48,7 @@ abstract class EntityController extends ApiController
     {
         $collection = $this->getAllEntities();
         $collection = $this->getWithNested($collection, $request);
+        $this->authorize($collection);
 
         return $this->getResponse()
             ->transformer($this->getTransformer())
@@ -59,12 +62,13 @@ abstract class EntityController extends ApiController
         $offset = $rangeRequest->isGetLast() ? $totalCount - $limit : $rangeRequest->getOffset();
 
         if ($request->has('q')) {
-            $collection = $this->searchAllEntities($request->query('q'), $limit, $offset, $totalCount);
+            $collection = $this->searchAllEntities(base64_decode($request->query('q')), $limit, $offset, $totalCount);
         } else {
             $collection = $this->getAllEntities($limit, $offset);
         }
 
         $collection = $this->getWithNested($collection, $request);
+        $this->authorize($collection);
 
         return $this->getResponse()
             ->transformer($this->getTransformer())
@@ -82,6 +86,7 @@ abstract class EntityController extends ApiController
     {
         $model = $this->findOrFailEntity($id);
         $model = $this->getWithNested($model, $request);
+        $this->authorize($model);
 
         return $this->getResponse()
             ->transformer($this->getTransformer())
@@ -97,8 +102,9 @@ abstract class EntityController extends ApiController
     public function postOne(Request $request)
     {
         $model = $this->getModel()->newInstance();
-        $this->validateRequest($request->all(), $this->getValidationRules());
-        $model->fill($request->all());
+        $this->validateRequest($request->json()->all(), $this->getValidationRules());
+        $model->fill($request->json()->all());
+        $this->authorize($model);
         $model->save();
 
         return $this->getResponse()
@@ -116,12 +122,12 @@ abstract class EntityController extends ApiController
     public function putOne(Request $request, $id)
     {
         $this->checkEntityIdMatchesRoute($request, $id, $this->getModel());
-
         $model = $this->findOrNewEntity($id);
 
-        $this->validateRequest($request->all(), $this->getValidationRules());
+        $this->validateRequest($request->json()->all(), $this->getValidationRules());
 
-        $model->fill($request->all());
+        $model->fill($request->json()->all());
+        $this->authorize($model);
         $model->save($this->isLocalised($request));
 
         return $this->getResponse()
@@ -137,16 +143,19 @@ abstract class EntityController extends ApiController
      */
     public function putMany(Request $request)
     {
-        $requestCollection = $request->all();
+        $requestCollection = $request->json()->all();
 
         $this->validateRequestCollection($requestCollection, $this->getValidationRules());
         $existingModels = $this->findCollection($requestCollection);
 
         $modelCollection = $this->getModel()
-            ->hydrateRequestCollection($requestCollection, $existingModels)
-            ->each(function (BaseModel $model) use ($request) {
-                return $model->save($this->isLocalised($request));
-            });
+            ->hydrateRequestCollection($requestCollection, $existingModels);
+
+        $this->authorize($modelCollection);
+
+        $modelCollection->each(function (BaseModel $model) use ($request) {
+            return $model->save($this->isLocalised($request));
+        });
 
         return $this->getResponse()
             ->transformer($this->getTransformer())
@@ -166,9 +175,10 @@ abstract class EntityController extends ApiController
 
         $model = $this->findOrFailEntity($id);
 
-        $this->validateRequest($request->all(), $this->getValidationRules(), true);
+        $this->validateRequest($request->json()->all(), $this->getValidationRules(), true);
 
-        $model->fill($request->all());
+        $model->fill($request->json()->all());
+        $this->authorize($model);
         $model->save();
 
         return $this->getResponse()->noContent();
@@ -182,15 +192,18 @@ abstract class EntityController extends ApiController
      */
     public function patchMany(Request $request)
     {
-        $requestCollection = $request->all();
+        $requestCollection = $request->json()->all();
 
         $this->validateRequestCollection($requestCollection, $this->getValidationRules(), true);
 
         $existingModels = $this->findOrFailCollection($requestCollection);
 
-        $this->getModel()
-            ->hydrateRequestCollection($requestCollection, $existingModels)
-            ->each(function (BaseModel $model) {
+        $modelsCollection = $this->getModel()
+            ->hydrateRequestCollection($requestCollection, $existingModels);
+
+        $this->authorize($existingModels);
+
+        $modelsCollection->each(function (BaseModel $model) {
                 return $model->save();
             });
 
@@ -205,7 +218,11 @@ abstract class EntityController extends ApiController
      */
     public function deleteOne($id)
     {
-        $this->findOrFailEntity($id)->delete();
+        $entity = $this->findOrFailEntity($id);
+
+        $this->authorize($entity);
+
+        $entity->delete();
 
         return $this->getResponse()->noContent();
     }
@@ -218,10 +235,13 @@ abstract class EntityController extends ApiController
      */
     public function deleteMany(Request $request)
     {
-        $requestCollection = $request->all();
+        $requestCollection = $request->json()->all();
 
-        $this->findOrFailCollection($requestCollection)
-            ->each(function (BaseModel $model) {
+        $modelsCollection = $this->findOrFailCollection($requestCollection);
+
+        $this->authorize($modelsCollection);
+
+        $modelsCollection->each(function (BaseModel $model) {
                 $model->delete();
             });
 
@@ -273,25 +293,35 @@ abstract class EntityController extends ApiController
     }
 
     /**
-     * @param $queryString
+     * @param $query
      * @param null $limit
      * @param null $offset
      * @param null $totalCount
      * @return \Elasticquent\ElasticquentResultCollection
      */
-    protected function searchAllEntities($queryString, $limit = null, $offset = null, &$totalCount = null)
+    protected function searchAllEntities($query, $limit = null, $offset = null, &$totalCount = null)
     {
         /* @var ElasticquentTrait $model */
         $model = $this->getModel();
 
-        $searchResults = $model->searchByQuery([
-            'match_phrase' => [
-                '_all' => $queryString,
-            ],
-        ], null, null, $limit, $offset);
+        $queryArray = json_decode($query, true);
+
+        if (is_array($queryArray)) { // Complex query
+            $searchResults = $model->complexSearch([
+                'index' => $model->getIndexName(),
+                'type' => $model->getTypeName(),
+                'body' => $this->translateQuery($queryArray),
+            ]);
+        } else {
+            $searchResults = $model->searchByQuery([
+                'match_phrase' => [
+                    '_all' => $query,
+                ],
+            ], null, null, $limit, $offset);
+        }
 
         if ($searchResults->totalHits() === 0) {
-            throw new NotFoundHttpException(sprintf('No results found with query `%s` for model `%s`', $queryString, get_class($model)));
+            throw new NotFoundHttpException(sprintf('No results found for model `%s`', get_class($model)));
         }
 
         if (isset($totalCount) && $searchResults->totalHits() < $totalCount) {
@@ -299,6 +329,73 @@ abstract class EntityController extends ApiController
         }
 
         return $searchResults;
+    }
+
+    /**
+     * Takes a query and translates it into a query that elastic search understands.
+     *
+     * Expect query to be in form, e.g.:
+     * {
+     *      _all: [ "stringA", "stringB" ],
+     *      someField: [ "stringA" ],
+     *      _nestedEntity: { key: [ "stringA", "stringB" ]
+     * }
+     *
+     * Notes:
+     * - Empty values will be removed from search completely.
+     * - You must pass an array, even if it only contains 1 string.
+     *
+     * @param $query
+     * @return mixed
+     */
+    private function translateQuery($query)
+    {
+        $processedQuery['query']['bool']['must'] = [];
+
+        foreach ($query as $key => $value) {
+            if (Str::startsWith($key, '_') && $key != '_all') { // Nested entity
+                reset($value);
+                $fieldKey = key($value);
+
+                foreach ($value[$fieldKey] as $fieldValue) {
+                    if (! empty($fieldValue)) {
+                        $snakeKey = snake_case(substr($key, 1));
+                        array_push($processedQuery['query']['bool']['must'],
+                            [
+                                'nested' => [
+                                    'path' => $snakeKey,
+                                    'query' => [
+                                        'bool' => [
+                                            'must' => [
+                                                'match' => [
+                                                    $snakeKey.'.'.snake_case($fieldKey) => $fieldValue,
+                                                ],
+                                            ],
+                                        ],
+                                    ],
+                                ],
+                            ]
+                        );
+                    }
+                }
+            } else {
+                foreach ($value as $matchValue) {
+                    if (! empty($matchValue)) {
+                        array_push($processedQuery['query']['bool']['must'], [
+                            'match' => [
+                                snake_case($key) => $matchValue,
+                            ],
+                        ]);
+                    }
+                }
+            }
+        }
+
+        if (empty($processedQuery['query']['bool']['must'])) { // No search params have been supplied, match all
+            return ['query' => ['match_all' => []]];
+        }
+
+        return $processedQuery;
     }
 
     /**
